Improve installer

+++ Modules-Installer.ps1
@@ -18,10 +18,19 @@ process
         {
             ForEach (${Module} in @(Get-ChildItem -Directory "${SourcePath}"))
             {
-                Copy-Item -LiteralPath "${SourcePath}\${Module}" -Destination "${DestinationPath}" -Force -Recurse | Out-Null
-                Import-Module -Name ${Module}
+                if (-not(Test-Path -Path ${env:SystemRoot}\System32\WindowsPowerShell\v1.0\Modules\${Module}))
+                {
+                    Write-Host "Installing module ${Module} to ${DestinationPath}"
+                    Copy-Item -LiteralPath "${SourcePath}\${Module}" -Destination "${DestinationPath}" -Recurse | Out-Null
+                    Import-Module -Name "${Module}" -Global -Scope Global
+                }
+                else
+                {
+                    Write-Host "Updating module ${Module} located at ${DestinationPath}"
+                    Copy-Item -LiteralPath "${SourcePath}\${Module}" -Destination "${DestinationPath}" -Force -Recurse | Out-Null
+                    Import-Module -Name "${Module}" -Force -Global -Scope Global
+                }
             }
-
             Write-Host -NoNewLine 'Press any key to continue...'
             $null = $Host.UI.RawUI.ReadKey('NoEcho,IncludeKeyDown')
         }
@@ -31,15 +40,15 @@ process
             {
                 if (Test-Path -Path "${env:SystemRoot}\System32\WindowsPowerShell\v1.0\Modules\${Module}")
                 {
+                    Write-Host "Removing module ${Module} from ${DestinationPath}"
                     Remove-Item -Path "${env:SystemRoot}\System32\WindowsPowerShell\v1.0\Modules\${Module}" -Force -Recurse | Out-Null
-                    Remove-Module -Name "${Module}"
+                    Remove-Module -Name "${Module}" -Force
                 }
                 else
                 {
-                    Write-Warning -Message "The module '${Module}' is not installed"
+                    Write-Error -Message "The module '${Module}' is not installed" -Category NotInstalled
                 }
             }
-
             Write-Host -NoNewLine 'Press any key to continue...'
             $null = $Host.UI.RawUI.ReadKey('NoEcho,IncludeKeyDown')
         }
